Move functions to I18n
